ai 모드 구현 -> 점수 차이, 끝나는 순간, ai의 점수 계산 방식 수정 필요
