+++ YamiTetris/tetris/AITetris.py
@@ -1,13 +1,12 @@
+import pygame, sys, time
+
+from ai_Board import *
 import random
 from ai import Ai
-import pygame, sys
-from gui import Gui
 
-# The configuration
-cell_size = 25
-cols = 10
-rows = 18
 
+
+#            R    G    B
 BLACK = (0, 0, 0)
 RED = (225, 13, 27)
 GREEN = (98, 190, 68)
@@ -16,12 +15,17 @@ ORANGE = (253, 189, 53)
 YELLOW = (246, 227, 90)
 PINK = (242, 64, 235)
 CYON = (70, 230, 210)
-GRAY = (23, 23, 23)
+GRAY = (26, 26, 26)
 WHITE = (255, 255, 255)
-
 colors = [BLACK, RED, GREEN, BLUE, ORANGE, YELLOW, PINK, CYON, GRAY]
 
-# Define the shapes of the single parts
+cell_size = 25
+cols = 10
+rows = 18
+base_width = 350
+base_height = 450
+fps = 200
+
 ai_tetris_shapes = [
     [[1, 1, 1],
      [0, 1, 0]],
@@ -86,13 +90,21 @@ def ai_new_board():
     return ai_board
 
 
+
+
+
 class AITetris(object):
+    #생성자
     def __init__(self, seed):
+        self.screen = pygame.display.set_mode((base_width*2, base_height)) # 고정 크기의 창을 만들어준다.  350 450
+        self.clock = pygame.time.Clock()
+        self.board = Board(self.screen)
+        self.music_on_off = True
+        self.check_reset = True
         random.seed(seed)
         self.width = cell_size * (cols + 4)  # 게임의 너비
         self.height = cell_size * rows
         self.next_stone = ai_tetris_shapes[random.randint(0, len(ai_tetris_shapes) - 1)]  # 다음 블럭 랜덤으로 고르기 0~6 사이의 랜덤 숫자를 통해 고르기
-        self.gui = Gui()
         self.ai_init_game()
 
     def ai_new_stone(self):
@@ -104,6 +116,7 @@ class AITetris(object):
         if ai_check_collision(self.ai_board, self.stone, (self.stone_x, self.stone_y)):
             self.gameover = True  # 블럭이 부딪히는 판단, 새로 생성한 블럭이 벽에 부딪히면은 게임 종료
 
+    #초기화 부분
     def ai_init_game(self):
         self.ai_board = ai_new_board()  # 새로운 게임 보드 생성
         self.ai_new_stone()  # 새로운 블럭 생성
@@ -111,7 +124,8 @@ class AITetris(object):
         self.ai_score = 0  # 시작 스코어
         self.ai_lines = 0  # 지운 라인의 개수
 
-    # 라인 지운 것에 대한 점수 및 지운 개수 추가
+        # 라인 지운 것에 대한 점수 및 지운 개수 추가
+
     def ai_add_cl_lines(self, n):
         linescores = [0, 40, 100, 300, 1200]
         self.ai_lines += n  # 지운 개수 추가하기
@@ -119,7 +133,8 @@ class AITetris(object):
         if self.ai_lines >= self.ai_level * 6:  # 지운 라인의 개수 >= 레벨*6 이되면 레벨 +1 하기
             self.ai_level += 1
 
-    # 블럭을 delta_x 만큼 움직이기
+        # 블럭을 delta_x 만큼 움직이기
+
     def ai_move(self, delta_x):
         if not self.gameover and not self.paused:  # 게임 종료, 정지 상태가 아니라면
             new_x = self.stone_x + delta_x  # 새로운 x 좌표는   기존의 stone의 x좌표 + 이동좌표수
@@ -148,7 +163,6 @@ class AITetris(object):
                 return True  # 게임 종료 상태가 아니 었다면,  true 반환
         return False  # 게임이 정지, 종료 상태 였다면 false 반환
 
-    # 블럭 회전
     def rotate_stone(self):
         if not self.gameover and not self.paused:
             ai_new_stone = ai_rotate_clockwise(self.stone)
@@ -157,18 +171,15 @@ class AITetris(object):
                                    (self.stone_x, self.stone_y)):
                 self.stone = ai_new_stone
 
-    # 게임 시작
     def ai_start_game(self):
         if self.gameover:  # 게임이 종료된 상태라면
             self.ai_init_game()  # 초기화 진행하기
             self.gameover = False  # 게임 종료는 false로 바꿔주기
 
-    # 게임 종료
     def ai_quit(self):
         pygame.display.update()  # 게임 화면 업데이트해주기
         sys.exit()  # 시스템 종료 하기
 
-    # 키에 해당하는 동작 지정해 주기 (ai가 동작 할 것 지정해 줄때 사용)
     def ai_executes_moves(self, ai_moves):
         ai_key_actions = {
             'LEFT': lambda: self.ai_move(-1),
@@ -182,32 +193,105 @@ class AITetris(object):
         for ai_action in ai_moves:
             ai_key_actions[ai_action]()
 
-    # 게임 시작 하기  (학습되어 있는 weight값 받아오기)
+
+    ###############################################################################
+    #각 키를 누를떄 실행되는 method
+    def handle_key(self, event_key):
+        if event_key == K_DOWN or event_key == K_s:
+            self.board.drop_piece()
+        elif event_key == K_LEFT or event_key == K_a:
+            self.board.move_piece(dx=-1, dy=0)
+        elif event_key == K_RIGHT or event_key == K_d:
+            self.board.move_piece(dx=1, dy=0)
+        elif event_key == K_UP or event_key == K_w:
+            self.board.rotate_piece()
+        elif event_key == K_SPACE:
+            self.board.full_drop_piece()
+        elif event_key == K_q: #스킬 부분
+            self.board.ultimate()
+        elif event_key == K_m: # 소리 설정
+            self.music_on_off = not self.music_on_off
+            if self.music_on_off:
+                pygame.mixer.music.play(-1, 0.0)
+            else:
+                pygame.mixer.music.stop()
+
+    #가장 높은 점수 불러 오는 부분
+    def HighScore(self):
+        try:
+            f = open('assets/ai_save.txt', 'r')
+            l = f.read()
+            f.close()
+            if int(l) < self.board.score:
+                h_s = self.board.score
+                f = open('assets/ai_save.txt', 'w')
+                f.write(str(self.board.score))
+                f.close()
+            else:
+                h_s = l
+            self.board.HS(str(h_s))
+        except:
+            f = open('assets/ai_save.txt', 'w')
+            f.write(str(self.board.score))
+            f.close()
+            self.board.HS(str(self.board.score))
+
+    #실행하기
     def run(self, weights):
+        pygame.init()
+        icon = pygame.image.load('assets/images/icon.PNG')  # png -> PNG로 수정
+        pygame.display.set_icon(icon)
+        pygame.display.set_caption('Tetris')
+
+        self.board.level_speed()
+
+        #start_sound = pygame.mixer.Sound('assets/sounds/Start.wav')
+        #start_sound.play()
+        #bgm = pygame.mixer.music.load('assets/sounds/bensound-ukulele.mp3')  # (기존 파일은 소리가 안남) 다른 mp3 파일은 소리 난다. 게임진행 bgm변경
         self.gameover = False
         self.paused = False
+        delay = 150
+        interval = 100
+        pygame.key.set_repeat(delay, interval)
 
-        # dont_burn_my_cpu = pygame.time.Clock()
-        while 1:
+        while True:
 
-            self.gui.update(self)
+            if self.check_reset:
+                self.board.newGame()
+                self.ai_init_game()
+                self.check_reset = False
+                #pygame.mixer.music.play(-1, 0.0)  ## 수정 필요 오류 나서 일단 빼둠
 
-            if self.gameover:
-                return self.ai_lines * 1000
+            #게임 종료시 어떻게 할건가
+            if self.board.game_over() or self.board.score  < self.ai_score:
+                self.screen.fill(BLACK) #게임 오버 배경 색
+                #pygame.mixer.music.stop() #음악 멈추기     오류나서 일단 뺴
+                #pygame.mixer.music.stop() #음악 멈추기     오류나서 일단 뺴
+                self.board.GameOver()  #게임 오버 보드 불러오기
+                self.HighScore()          #하이스코어 표기
+                self.check_reset = True
+                self.board.init_board()
 
             Ai.choose(self.ai_board, self.stone, self.next_stone, self.stone_x, weights, self)
 
-            for event in pygame.event.get():
-                if event.type == pygame.USEREVENT + 1:
+            for event in pygame.event.get(): #게임진행중 - event는 키보드 누를떄 특정 동작 수할떄 발생
+                if event.type == QUIT: #종류 이벤트가 발생한 경우
+                    pygame.quit() #모든 호출 종
+                    sys.exit() #게임을 종료한다ㅏ.
+                elif event.type == KEYUP and event.key == K_p: # 일시 정지 버튼 누르면
+                    self.screen.fill(BLACK)         #일시 정지 화면
+                    #pygame.mixer.music.stop()       #일시 정지 노래 중둠    오류나서 일단 뺴
+                    self.board.pause()
+                    #pygame.mixer.music.play(-1, 0.0)
+                elif event.type == KEYDOWN: #키보드를 누르면
+                    self.handle_key(event.key) #handle 메소드 실행
+                elif event.type == pygame.USEREVENT or event.type == pygame.USEREVENT+1:
+                    self.board.drop_piece()
                     self.ai_drop(False)
-                elif event.type == pygame.QUIT:
-                    self.ai_quit()
-                # elif event.type == pygame.KEYDOWN:
-                #    if event.key == eval("pygame.K_p"):
-                #      self.toggle_pause()
 
-            # dont_burn_my_cpu.tick(maxfps)
-
+            self.board.draw(self,True)
+            pygame.display.update() #이게 나오면 구현 시
+            self.clock.tick(fps) # 초당 프레임 관련
 
 if __name__ == '__main__':
     weights = [3.39357083734159515, -1.8961941343266449, -5.107694873375318, -3.6314963941589093, -2.9262681134021786,
@@ -218,4 +302,3 @@ if __name__ == '__main__':
                -5.3416512085013395, -4.072687054171711, -5.936652569831475, -2.3140398163110643, -4.842883337741306,
                17.677262456993276, -4.42668539845469, -6.8954976464473585, 4.481308299774875]  # 21755 lignes
     AITetris(4).run(weights)
-
