+++ YamiTetris/tetris/Menu.py
@@ -2,6 +2,8 @@ import pygame
 import pygame_menu
 from Board import *
 from Tetris import *
+from AITetris import AITetris
+
 pygame.init()
 surface=pygame.display.set_mode((800,600))
 
@@ -49,7 +51,18 @@ def start_the_Twohands():
 
 def start_the_Ai():
     ## ai 모드 만들면 여기다 실행 코드 넣으세요
-    pass
+    weights = [3.39357083734159515, -1.8961941343266449, -5.107694873375318, -3.6314963941589093,
+               -2.9262681134021786,
+               -2.146136640641482, -7.204192964669836, -3.476853402227247, -6.813002842291903, 4.152001386170861,
+               -21.131715861293525, -10.181622180279133, -5.351108175564556, -2.6888972099986956,
+               -2.684925769670947,
+               -4.504495386829769, -7.4527302422826, -6.3489634714511505, -4.701455626343827, -10.502314845278828,
+               0.6969259450910086, -4.483319180395864, -2.471375907554622, -6.245643268054767, -1.899364785170105,
+               -5.3416512085013395, -4.072687054171711, -5.936652569831475, -2.3140398163110643, -4.842883337741306,
+               17.677262456993276, -4.42668539845469, -6.8954976464473585, 4.481308299774875]  # 21755 lignes
+    if __name__ == '__main__':
+         AITetris(4).run(weights)
+
 
 def show_the_rank():
     ## 일반게임 랭크 보여주기
