+++ YamiTetris/tetris/assets/ai_save.txt
@@ -1,0 +1,1 @@
+11
