+++ YamiTetris/tetris/assets/save.txt
@@ -1,1 +1,1 @@
-9
+68
